Added bow

+++ Assets/Scripts/Weapon/WeaponData.cs
@@ -7,10 +7,9 @@ public class WeaponData : ScriptableObject, IWeaponable
 {
     [field: SerializeField] public WeaponType WeaponType { get; private set; }
 
-    [field: SerializeField] public int Damage { get; private set; }
+    [field: SerializeField] public float Damage { get; private set; }
 
     [field: SerializeField] public int AttackSpeed { get; private set; }//¬озможно, должно быть в классе игрока
 
     [field: SerializeField] public float AttackRadius { get; private set; }
-
 }
